+++ src/main/kotlin/com/xmartlabs/snapshotpublisher/task/ErrorTask.kt
@@ -1,9 +1,11 @@
 package com.xmartlabs.snapshotpublisher.task
 
 import org.gradle.api.DefaultTask
+import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.TaskAction
 
 open class ErrorTask : DefaultTask() {
+  @Input
   lateinit var message: String
 
   @Suppress("unused")